Exporting and unexporting TableResultListener while synchronized.
We're getting an exception on stop(), because the TableResultListener
itself is not serializable.  We expect it to be exported, and thus used
via its Proxy.  However, it is trying to serialize it, which seems to
indicate the TableResultListener is not exported.

This implementation only tries to remove the table listener when it is
flagged as exported.